Adding more frequent bandwidth checks
Now we calculate bandwidth while it's being reported and assess quality
before constructing download manifests.

This means that if you are on a laptop on a fast network, we retain that
bandwidth. If you switch to a slower network, this change means that
preemptive requests will down shift much faster.